Remove modal link from tables in print
